# apps/api/openapi.json
{
  "components": {
    "schemas": {
      "AccountStat": {
        "properties": {
          "account": {
            "title": "Account",
            "type": "string"
          },
          "avg_likes": {
            "title": "Avg Likes",
            "type": "number"
          },
          "avg_views": {
            "title": "Avg Views",
            "type": "number"
          },
          "posts": {
            "title": "Posts",
            "type": "integer"
          }
        },
        "required": [
          "account",
          "posts",
          "avg_likes",
          "avg_views"
        ],
        "title": "AccountStat",
        "type": "object"
      },
...
      "BillingUsageResponse": {
        "properties": {
          "credits": {
            "$ref": "#/components/schemas/CreditState"
          },
          "plan": {
            "title": "Plan",
            "type": "string"
          },
          "usage": {
            "$ref": "#/components/schemas/UsageCounts"
          }
        },
        "required": [
          "plan",
          "usage",
          "credits"
        ],
        "title": "BillingUsageResponse",
        "type": "object"
      },
...
      "CheckoutRequest": {
        "properties": {
          "price_id": {
            "title": "Price Id",
            "type": "string"
          }
        },
        "required": [
          "price_id"
        ],
        "title": "CheckoutRequest",
        "type": "object"
      },
      "CreditState": {
        "properties": {
          "monthly_left": {
            "title": "Monthly Left",
            "type": "integer"
          },
          "monthly_limit": {
            "title": "Monthly Limit",
            "type": "integer"
          },
          "prepaid_left": {
            "title": "Prepaid Left",
            "type": "integer"
          },
          "prepaid_limit": {
            "title": "Prepaid Limit",
            "type": "integer"
          }
        },
        "required": [
          "monthly_left",
          "prepaid_left",
          "monthly_limit",
          "prepaid_limit"
        ],
        "title": "CreditState",
        "type": "object"
      },
      "FeedResponse": {
        "properties": {
          "audio_url": {
            "default": "",
            "title": "Audio Url",
            "type": "string"
          },
          "is_demo": {
            "default": false,
            "title": "Is Demo",
            "type": "boolean"
          },
          "last_generation_time": {
            "default": "",
            "title": "Last Generation Time",
            "type": "string"
          },
          "last_generation_time_local": {
            "default": "",
            "title": "Last Generation Time Local",
            "type": "string"
          },
          "raw_data_sources": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Raw Data Sources",
            "type": "array"
          },
          "summary_html": {
            "default": "",
            "title": "Summary Html",
            "type": "string"
          },
          "timezone": {
            "default": "America/New_York",
            "title": "Timezone",
            "type": "string"
          }
        },
        "title": "FeedResponse",
        "type": "object"
      },
      "GenerationCostResponse": {
        "properties": {
          "blockers": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Blockers",
            "type": "array"
          },
          "can_generate": {
            "title": "Can Generate",
            "type": "boolean"
          },
          "cost": {
            "title": "Cost",
            "type": "integer"
          },
          "credits": {
            "$ref": "#/components/schemas/CreditState"
          }
        },
        "required": [
          "cost",
          "credits",
          "can_generate"
        ],
        "title": "GenerationCostResponse",
        "type": "object"
      },
      "GenerationStatus": {
        "properties": {
          "current_stage": {
            "default": "",
            "title": "Current Stage",
            "type": "string"
          },
          "end_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "stages_completed": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Stages Completed",
            "type": "array"
          },
          "start_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time"
          },
          "status": {
            "default": "",
            "title": "Status",
            "type": "string"
          }
        },
        "title": "GenerationStatus",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Message": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "Message",
        "type": "object"
      },
...
      "SourceDataResponse": {
        "properties": {
          "per_account": {
            "items": {
              "$ref": "#/components/schemas/AccountStat"
            },
            "title": "Per Account",
            "type": "array"
          },
          "rows": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Rows",
            "type": "array"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/TimelinePoint"
            },
            "title": "Timeline",
            "type": "array"
          },
          "total_likes": {
            "title": "Total Likes",
            "type": "integer"
          },
          "total_posts": {
            "title": "Total Posts",
            "type": "integer"
          },
          "total_views": {
            "title": "Total Views",
            "type": "integer"
          }
        },
        "required": [
          "total_posts",
          "total_likes",
          "total_views",
          "per_account",
          "timeline",
          "rows"
        ],
        "title": "SourceDataResponse",
        "type": "object"
      },
      "StartGenerationRequest": {
        "properties": {
          "fetch_latest": {
            "default": true,
            "title": "Fetch Latest",
            "type": "boolean"
          },
          "prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt"
          },
          "skip_audio": {
            "default": false,
            "title": "Skip Audio",
            "type": "boolean"
          },
          "skip_email": {
            "default": false,
            "title": "Skip Email",
            "type": "boolean"
          }
        },
        "title": "StartGenerationRequest",
        "type": "object"
      },
      "TimelinePoint": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "time": {
            "title": "Time",
            "type": "string"
          }
        },
        "required": [
          "time",
          "count"
        ],
        "title": "TimelinePoint",
        "type": "object"
      },
...
      "UrlResponse": {
        "properties": {
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "UrlResponse",
        "type": "object"
      },
      "UsageCounts": {
        "properties": {
          "n_chat_messages": {
            "default": 0,
            "title": "N Chat Messages",
            "type": "integer"
          },
          "n_followers_scraped": {
            "default": 0,
            "title": "N Followers Scraped",
            "type": "integer"
          },
          "n_generations": {
            "default": 0,
            "title": "N Generations",
            "type": "integer"
          },
          "n_newsletters_sent": {
            "default": 0,
            "title": "N Newsletters Sent",
            "type": "integer"
          }
        },
        "title": "UsageCounts",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "API wrapping the FeedTLDR pipeline (scrape, summarize, narrate, email). Authenticated endpoints expect a Firebase ID token as a Bearer token.",
    "title": "FeedTLDR API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/billing/checkout": {
      "post": {
        "operationId": "create_checkout_v1_billing_checkout_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UrlResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Checkout",
        "tags": [
          "billing"
        ]
      }
    },
...
    "/v1/billing/portal": {
      "post": {
        "operationId": "create_portal_v1_billing_portal_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UrlResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Portal",
        "tags": [
          "billing"
        ]
      }
    },
    "/v1/billing/usage": {
      "get": {
        "operationId": "get_usage_v1_billing_usage_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingUsageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Usage",
        "tags": [
          "billing"
        ]
      }
    },
...
    "/v1/feed": {
      "get": {
        "operationId": "get_feed_v1_feed_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Feed",
        "tags": [
          "feed"
        ]
      }
    },
    "/v1/feed/source-data": {
      "get": {
        "operationId": "get_source_data_v1_feed_source_data_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDataResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Source Data",
        "tags": [
          "feed"
        ]
      }
    },
    "/v1/generations": {
      "post": {
        "operationId": "start_generation_v1_generations_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartGenerationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Generation",
        "tags": [
          "generations"
        ]
      }
    },
    "/v1/generations/cost": {
      "get": {
        "operationId": "generation_cost_v1_generations_cost_get",
        "parameters": [
          {
            "in": "query",
            "name": "fetch_latest",
            "required": false,
            "schema": {
              "default": true,
              "title": "Fetch Latest",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "skip_audio",
            "required": false,
            "schema": {
              "default": false,
              "title": "Skip Audio",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationCostResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generation Cost",
        "tags": [
          "generations"
        ]
      }
    },
    "/v1/generations/status": {
      "get": {
        "description": "The polling seam: reads the same pipeline_status document the pipeline\nwrites stage transitions to (poll every 5-8s, like the legacy dialog).",
        "operationId": "generation_status_v1_generations_status_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generation Status",
        "tags": [
          "generations"
        ]
      }
    },
...
  }
}